old-tests/submission/Microsoft/forms/formnovalidate.html

<!DOCTYPE html>
<html><head>
        <title>formnovalidate on button element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="content-type">
        <meta content="formnovalidate" name="description">
    <link href="http://www.w3.org/TR/html5/attributes-common-to-form-controls.html#dom-fs-formnovalidate" rel="help">
</head>
    <body>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <div id="log"></div>
    
        <form style="display : none">
          <label for="input1">Required<input id="input1" name="input1" required="" type="text"></label>
          <br>
          <button formnovalidate="" type="submit">Submit No Validate</button>
          <br>
          <button type="submit">Submit Validate</button>
          <br>
          <input formnovalidate="" type="submit" value="Submit No Validate">
          <br>
          <input type="submit" value="Submit Validate">
        </form>

        <script type="text/javascript">
            test(function() {assert_true(document.getElementsByTagName("button")[0].formNoValidate)}, "formNoValidate returns true when specified as an attribute on a button element");
            test(function() {assert_false(document.getElementsByTagName("button")[1].formNoValidate)}, "formNoValidate returns false when not specified as an attribute on a button element");
            test(function() {assert_true(document.getElementsByTagName("input")[1].formNoValidate)}, "formNoValidate returns true when specified as an attribute on a input element");
            test(function() {assert_false(document.getElementsByTagName("input")[2].formNoValidate)}, "formNoValidate returns false when not specified as an attribute on a input element");

            var nSubmitEventCalled = 0;
            var nInvalidEventCalled = 0;
            
            function handleSubmit(e){nSubmitEventCalled++;e.preventDefault();}
            function handleInvalid(e){nInvalidEventCalled++;e.preventDefault();}
            
            var testElem = document.getElementsByTagName("input")[1];
            var formElem = document.getElementsByTagName("form")[0];

            formElem.addEventListener("submit", handleSubmit, false);
            testElem.click();

            test(function() {assert_equals(nSubmitEventCalled, 1)}, "click input element with formnovalidate attribute");

            nSubmitEventCalled = 0;
            
            testElem = document.getElementsByTagName("input")[2];
            requiredElem = document.getElementsByTagName("input")[0];

            requiredElem.addEventListener("invalid", handleInvalid, false);

            testElem.click();
            
            test(function() {assert_equals(nSubmitEventCalled, 0)}, "click input element without formnovalidate attribute");
            test(function() {assert_equals(nInvalidEventCalled, 1)}, "click input element without formnovalidate attribute");
            
        </script>
    
</body></html>
